Move function to write headers into spreadsheet.js
